# server/src/main/resources/openapi/caudal-api.yaml
openapi: 3.0.3
info:
  title: Caudal API
  version: 0.1.0
  description: |
    Caudal is a **temporal relevance memory for AI agents**. It tells an agent
    what matters *right now*.

    Agents emit interaction events (a user mentioned a topic, a tool was used,
    a document was retrieved). Caudal continuously reinforces recent activity,
    naturally forgets what is no longer relevant, and exposes a ranked
    "attention signal" that agents can query before every decision.

    ## Core concepts

    | Concept | Description |
    |---------|-------------|
    | **Space** | An isolated memory namespace. Typically one per user, session, or tenant. Spaces are created implicitly the first time an event references them. |
    | **Entity** | Any string identifier that participates in events (`user:123`, `topic:cars`, `tool:search`). Entities are graph nodes. |
    | **Event** | A directed, weighted interaction between two entities (`src` → `dst`). Events reinforce the relationship and both entities' relevance. |
    | **Focus** | A ranked list of the most relevant entities in a space *right now*, computed from decayed cumulative interaction scores. |
    | **Next hop** | Given a source entity, the ranked outgoing neighbors — what is most strongly associated with it at this moment. |
    | **Pathway** | A multi-hop walk through the association graph. Pathways reveal indirect connections and chains of relevance useful for planning, recommendations, and explanations. |

...

...

...

...

...

...
servers:
...
paths:
...
  /pathways:
    post:
      operationId: pathways
      tags: [Queries]
      summary: Discover multi-hop pathways from a starting entity
      description: |
        Samples probabilistic walks through the association graph starting
        from the given entity. Returns the most frequently traversed paths
        and the most visited entities across all walks.

        Pathways reveal **indirect, multi-hop connections** that single-hop
        queries (`/next`) cannot surface. Use this for:

        - **Planning:** discover chains of relevant entities for multi-step
          agent plans.
        - **Recommendations:** find non-obvious connections between entities
          (e.g. user → topic → brand → product).
        - **Explanations:** generate human-readable relevance traces
          ("because you were looking at cars → Toyota → Yaris Cross").

        The `mode` parameter controls the trade-off between latency and
        thoroughness without exposing internal algorithm parameters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathwayRequest'
            example:
              space: "user:123"
              start: "user:123"
              k: 5
              mode: "balanced"
      responses:
        '200':
          description: |
            The discovered pathways and most-visited entities. Paths are
            ordered by descending probability (how frequently the walk
            followed that exact sequence). Top entities are ordered by
            visit frequency across all walks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathwayResponse'
              example:
                paths:
                  - nodes: ["user:123", "topic:car-buying", "brand:toyota", "model:yaris_cross"]
                    score: 0.34
                  - nodes: ["user:123", "topic:car-buying", "doc:comparison_guide"]
                    score: 0.28
                topEntities:
                  - id: "topic:car-buying"
                    score: 0.72
                  - id: "brand:toyota"
                    score: 0.51
                  - id: "model:yaris_cross"
                    score: 0.44
                asOf: "2026-02-28T10:06:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------
components:
...
  responses:
    BadRequest:
      description: |
        The request was malformed or failed validation. The `error` field
        contains a human-readable description of the problem.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "space: must not be blank"
...
  schemas:
...
    ScoredItem:
      type: object
      description: |
        An entity and its current relevance score.
      properties:
        id:
          type: string
          description: |
            The entity identifier, matching the `src` or `dst` values from
            previously ingested events.
          example: "topic:car-buying"
        score:
          type: number
          format: double
          description: |
            Current relevance score after time-based decay. Higher means
            more relevant right now. Scores are non-negative and decrease
            over time as interactions age. The absolute magnitude depends on
            event weights and recency; compare scores within the same
            response to determine relative importance.
          example: 0.83

    # === Pathways ===
    PathwayRequest:
      type: object
      description: |
        Request parameters for pathway discovery. Caudal samples
        probabilistic walks through the association graph and returns the
        most common paths and visited entities.
      required: [space, start]
      properties:
        space:
          type: string
          minLength: 1
          description: |
            The space to explore. Must match a space that has received
            events.
          example: "user:123"
        start:
          type: string
          minLength: 1
          description: |
            The entity to start all walks from. Typically the current user,
            active topic, or any entity whose indirect associations you want
            to discover.
          example: "user:123"
        k:
          type: integer
          format: int32
          default: 10
          description: |
            Maximum number of top paths and top entities to return.
            Defaults to `10` if omitted or non-positive.
          example: 5
        mode:
          type: string
          default: balanced
          enum: [fast, balanced, deep]
          description: |
            Controls the trade-off between response latency and exploration
            thoroughness.

            | Mode | Behavior |
            |------|----------|
            | `fast` | Fewer samples, shorter walks. Best for low-latency use cases where approximate results are acceptable. |
            | `balanced` | Moderate sampling depth. Good default for most agent workflows. Used when `mode` is omitted. |
            | `deep` | More samples, longer walks. Best for planning, root-cause analysis, or when discovering distant associations matters. |
    PathwayResponse:
      type: object
      description: |
        Results of pathway discovery, containing the most frequently
        traversed paths and the most visited entities across all sampled
        walks.
      properties:
        paths:
          type: array
          description: |
            The top-**k** most frequently traversed paths, ordered by
            descending score. Each path is an ordered sequence of entity
            identifiers representing one walk through the graph.
          items:
            $ref: '#/components/schemas/PathItem'
        topEntities:
          type: array
          description: |
            The top-**k** most frequently visited entities across all
            walks, ordered by descending visit frequency. Useful for
            identifying the most important entities reachable from the
            starting point, regardless of the specific path taken.
          items:
            $ref: '#/components/schemas/ScoredItem'
        asOf:
          type: string
          description: |
            ISO-8601 instant of the current server time bucket at which
            the pathway analysis was performed.
          example: "2026-02-28T10:06:00Z"
    PathItem:
      type: object
      description: |
        A single discovered pathway — an ordered sequence of entities
        representing one frequently traversed walk through the
        association graph.
      properties:
        nodes:
          type: array
          description: |
            Ordered list of entity identifiers along the path. The first
            element is always the `start` entity from the request. Each
            subsequent element was reached by following the highest-weight
            outgoing edge probabilistically.
          items:
            type: string
          example: ["user:123", "topic:car-buying", "brand:toyota", "model:yaris_cross"]
        score:
          type: number
          format: double
          description: |
            The fraction of sampled walks that followed this exact path.
            Ranges from `0.0` to `1.0`. A score of `0.34` means 34% of
            all walks traversed this specific sequence of entities.
          example: 0.34

    # === Errors ===
    ErrorResponse:
      type: object
      description: |
        Standard error envelope returned for all `4xx` responses.
      properties:
        error:
          type: string
          description: |
            A human-readable error message describing what went wrong.
            For validation errors, this includes the field name and
            constraint that failed.
          example: "space: must not be blank"

  # ----- Security -----
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Pass the server-configured API key as a
        Bearer token. The key can also be sent with the `Token` prefix
        (i.e. `Authorization: Token <key>`).
security:
  - BearerAuth: []